requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==25.3.0
backtesting==0.6.4
beautifulsoup4==4.13.4
bokeh==3.7.3
certifi==2025.4.26
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.4.2
contourpy==1.3.2
curl_cffi==0.11.1
cycler==0.12.1
discord-protos==0.0.2
discord.py-self @ file:///home/william/discord.py-self
filelock==3.18.0
fonttools==4.58.4
frozendict==2.4.6
fsspec==2025.3.2
greenlet==3.2.4
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
multidict==6.4.4
multitasking==0.0.12
mysql-connector-python==9.4.0
narwhals==1.41.0
networkx==3.4.2
numpy==2.2.5
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
packaging==25.0
pandas==2.2.3
peewee==3.18.2
pillow==11.2.1
platformdirs==4.3.8
propcache==0.3.1
protobuf==6.31.1
pycparser==2.22
pyparsing==3.2.3
PyQt6==6.9.1
PyQt6-Qt6==6.9.1
PyQt6_sip==13.10.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
scikit-learn==1.7.0
scipy==1.15.3
setuptools==80.4.0
six==1.17.0
soupsieve==2.7
SQLAlchemy==2.0.42
sympy==1.14.0
ta-lib==0.6.3
tabulate==0.9.0
schwab-py==1.5.1
threadpoolctl==3.6.0
torch==2.7.0
tornado==6.5.1
triton==3.3.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
websockets==15.0.1
xgboost==3.0.2
xyzservices==2025.4.0
yarl==1.20.0
yfinance==0.2.65
fastapi==0.115.12
uvicorn[standard]==0.34.0
apscheduler==3.11.0
